+++ common/src/lib.rs
@@ -1,24 +1,48 @@
-/// The mouse movement data in pixels in one update.
-#[derive(Default, Debug, Clone, Copy)]
-pub struct MouseMove {
-  pub x: i8,
-  pub y: i8,
+#[derive(Default, Debug, Clone, Copy, PartialEq, Eq)]
+pub struct MouseButton(pub u8);
+
+impl MouseButton {
+  pub const MOUSE_LEFT_BUTTON: u8 = 1;
+  pub const MOUSE_RIGHT_BUTTON: u8 = 2;
+
+  pub fn left_button_down(&mut self) {
+    self.0 |= Self::MOUSE_LEFT_BUTTON;
+  }
+  pub fn is_left_button_down(&self) -> bool {
+    self.0 & Self::MOUSE_LEFT_BUTTON != 0
+  }
+  pub fn right_button_down(&mut self) {
+    self.0 |= Self::MOUSE_RIGHT_BUTTON;
+  }
+  pub fn is_right_button_down(&self) -> bool {
+    self.0 & Self::MOUSE_RIGHT_BUTTON != 0
+  }
 }
 
-impl MouseMove {
+/// The mouse movement data in pixels in one update.
+#[derive(Default, Debug, Clone, Copy)]
+pub struct Mouse {
+  pub x: i8,
+  pub y: i8,
+  pub buttons: MouseButton,
+}
+
+impl Mouse {
   /// Serialize the mouse movement data into a buffer.
-  /// The buffer must be at least 2 bytes long.
+  /// The buffer must be at least 3 bytes long.
   pub fn serialize(&self, buf: &mut [u8]) {
     buf[0] = self.x as u8;
     buf[1] = self.y as u8;
+    buf[2] = self.buttons.0;
   }
 
   /// Deserialize the mouse movement data from a buffer.
-  /// The buffer must be at least 2 bytes long.
+  /// The buffer must be at least 3 bytes long.
   pub fn deserialize(buf: &[u8]) -> Self {
     Self {
       x: buf[0] as i8,
       y: buf[1] as i8,
+      buttons: MouseButton(buf[2]),
     }
   }
 }
@@ -27,8 +51,8 @@ impl MouseMove {
 pub enum Packet<Gamepad> {
   Timestamp(u64),
   Gamepad(Gamepad),
-  MouseMove(MouseMove),
-  GamepadAndMouseMove(Gamepad, MouseMove),
+  Mouse(Mouse),
+  GamepadAndMouse(Gamepad, Mouse),
 }
 
 pub const PACKET_FRAME_SIZE: usize = 16;