+++ deck/src/server.rs
@@ -64,11 +64,11 @@ impl<Gamepad: SerializableGamepad> SerializablePacket for Packet<Gamepad> {
         buf[0] = 1;
         gamepad.serialize(&mut buf[1..]);
       }
-      Packet::MouseMove(mouse) => {
+      Packet::Mouse(mouse) => {
         buf[0] = 2;
         mouse.serialize(&mut buf[1..]);
       }
-      Packet::GamepadAndMouseMove(gamepad, mouse) => {
+      Packet::GamepadAndMouse(gamepad, mouse) => {
         buf[0] = 3;
         gamepad.serialize(&mut buf[1..]);
         mouse.serialize(&mut buf[13..]);